fixed number of channels in gating
in the inital implementation, twice the in channels were used. However, this is not feasible. I assume this mistake was introduced by c&p from an external source that applied some transformation before, doubling the number of channels. My assumption is due to the subsequent conv. layer, which was expecting also double the channels.